Add github_sdk path to platforms
For k32w0, the SSBL binary is found in the SDK, which is
downloaded through west in the Matter repo. The scripts
assumes that the metadata repo is a submodule inside the Matter
repo, so it obtains the sdk path through relative path.

For k32w1, the NBU signed binary should be in the SDK, but it is
currently not the case. This should be updated in the future. In
the meantime, maintain a copy of the NBU in the binaries folder.

Signed-off-by: marius-alex-tache <marius.tache@nxp.com>